# docs/swagger.yaml
basePath: /api
definitions:
...
  controller.CreemPayRequest:
    properties:
      payment_method:
        type: string
      product_id:
        type: string
    type: object
...
  dto.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact: {}
  description: new-api AI 网关 —— 用户与账号相关接口(对外第三方)
  title: new-api 用户 API
  version: "1.0"
paths:
...
  /user/creem/pay:
    post:
      parameters:
      - description: 产品 ID 与支付方式
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.CreemPayRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Creem 充值下单
      tags:
      - 用户-充值
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"